+++ sdk/src/test/kotlin/dev/datastar/kotlin/sdk/SendEventTests.kt
@@ -4,7 +4,6 @@ import dev.datastar.kotlin.sdk.testcases.datastarTest
 import io.kotest.matchers.shouldBe
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.TestFactory
-import kotlin.properties.Delegates
 
 class SendEventTests {
     @TestFactory
@@ -136,28 +135,5 @@ class SendEventTests {
             
             
             """.trimIndent()
-    }
-}
-
-class TestResponse : Response {
-    var status by Delegates.notNull<Int>()
-    var headers: MutableMap<String, List<String>> = mutableMapOf()
-    var output: String = ""
-    var flushedCount = 0
-
-    override fun sendConnectionHeaders(
-        status: Int,
-        headers: Map<String, List<String>>,
-    ) {
-        this.status = status
-        this.headers += headers
-    }
-
-    override fun write(text: String) {
-        output += text
-    }
-
-    override fun flush() {
-        flushedCount++
     }
 }